seastore: seperate Journal interface from SegmentedJournal implementation
A subsequent PR will introduce a CircularBoundedJournal implementation
for fast nvme devices.

SegmentCleaner no longer needs a reference to Journal, so dispense with
the set_segment_provider machinery and simply pass it in the
constructor.

Move responsibility for finding the journal segments into the journal
itself.  This does mean that we check the segment headers on the journal
device twice, but that should be a neglible amount of overhead on mount.

SegmentCleaner::init_segments no longer needs to return Journal
segments, so merge with mount().
